api/requirements.txt
aiosqlite==0.17.0
anyio==3.6.2
asttokens==2.2.1
asyncpg==0.27.0
backcall==0.2.0
bcrypt==4.0.1
cffi==1.15.1
click==8.1.3
cryptography==40.0.2
databases==0.7.0
decorator==5.1.1
ecdsa==0.18.0
executing==1.2.0
fastapi==0.95.2
greenlet==2.0.2
h11==0.14.0
httptools==0.5.0
idna==3.4
iso8601==1.1.0
jedi==0.18.2
Jinja2==3.1.2
MarkupSafe==2.1.3
matplotlib-inline==0.1.6
parso==0.8.3
passlib==1.7.4
pexpect==4.8.0
pickleshare==0.7.5
prompt-toolkit==3.0.38
ptyprocess==0.7.0
pure-eval==0.2.2
pyasn1==0.5.0
pycparser==2.21
pydantic==1.10.7
Pygments==2.15.1
pypika-tortoise==0.1.6
python-dotenv
python-jose==3.3.0
pytz==2023.3
PyYAML==6.0
rsa==4.9
six==1.16.0
sniffio==1.3.0
SQLAlchemy==1.4.48
stack-data==0.6.2
starlette==0.27.0
tortoise-orm==0.19.3
traitlets==5.9.0
typing_extensions==4.5.0
uvicorn[standard]
uvloop==0.17.0
watchfiles==0.19.0
wcwidth==0.2.6
websockets==11.0.3
